requirement.txt
paddleocr>=2.6.0
paddlepaddle
openai
python-dotenv